% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{Board}
\alias{Board}
\title{R6 Class Representing a Board}
\description{
A container to house the minefield. but not the game state
}
\details{
A board is largly hidden to the player, they must discover it on their own, and maintain the state themselves
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ask}{\code{Board$ask()}}
\item \href{#method-submit}{\code{Board$submit()}}
\item \href{#method-new}{\code{Board$new()}}
\item \href{#method-print}{\code{Board$print()}}
\item \href{#method-clone}{\code{Board$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ask"></a>}}
\if{latex}{\out{\hypertarget{method-ask}{}}}
\subsection{Method \code{ask()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Board$ask(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Board$submit(candidate_minefield)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Board$new(height = 10, width = 10, mines = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Board$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Board$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
