% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{cellUI}
\alias{cellUI}
\title{Helper function to create cell element}
\usage{
cellUI(i, j, name, class = "")
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}

\item{name}{type of the cell}

\item{class}{class of the html element}
}
\description{
Helper function to create cell element
}
