% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{checkCell}
\alias{checkCell}
\title{Given a game, a state and i, j return a new state with the cell i, j and
recursively all its non-ambiguous neighbors checked}
\usage{
checkCell(game, state, i, j)
}
\arguments{
\item{game}{the game}

\item{state}{the state of the game}

\item{i}{the row coordinate}

\item{j}{the column coordinate}
}
\description{
Given a game, a state and i, j return a new state with the cell i, j and
recursively all its non-ambiguous neighbors checked
}
