% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{checkedCellUI}
\alias{checkedCellUI}
\title{Create checked cell element}
\usage{
checkedCellUI(i, j, nearby_mines)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}

\item{nearby_mines}{number of nearby mines}
}
\description{
Create checked cell element
}
