% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{countNearbyMines}
\alias{countNearbyMines}
\title{Count the number of nearby mines}
\usage{
countNearbyMines(mines, i, j)
}
\arguments{
\item{mines}{the mines grid}

\item{i}{the row coordinate}

\item{j}{the column coordinate}
}
\description{
Count the number of nearby mines
}
