% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createCheckedGrid}
\alias{createCheckedGrid}
\title{Create checked grid}
\usage{
createCheckedGrid(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Create checked grid
}
