% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createFlaggedGrid}
\alias{createFlaggedGrid}
\title{Create flagged grid}
\usage{
createFlaggedGrid(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Create flagged grid
}
