% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createInitialState}
\alias{createInitialState}
\title{Create initial state of the game}
\usage{
createInitialState(game)
}
\arguments{
\item{game}{the game}
}
\description{
Create initial state of the game
}
