% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createMinesGrid}
\alias{createMinesGrid}
\title{Create a mines grid}
\usage{
createMinesGrid(nrow, ncol, nmines)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{nmines}{number of mines}
}
\description{
Create a mines grid
}
