% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createNearbyMinesGrid}
\alias{createNearbyMinesGrid}
\title{Create a matrix that map i, j coordinates to the count of nearby mines}
\usage{
createNearbyMinesGrid(mines)
}
\arguments{
\item{mines}{the mines grid}
}
\description{
Create a matrix that map i, j coordinates to the count of nearby mines
}
