% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{createRevealedGrid}
\alias{createRevealedGrid}
\title{Create revealed grid}
\usage{
createRevealedGrid(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Create revealed grid
}
