% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{defeatGridUI}
\alias{defeatGridUI}
\title{Create defeat grid element}
\usage{
defeatGridUI(game, state)
}
\arguments{
\item{game}{the game to display}

\item{state}{the current state of the game}
}
\description{
Create defeat grid element
}
