% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{flagCellUI}
\alias{flagCellUI}
\title{Create flagged cell element}
\usage{
flagCellUI(i, j)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}
}
\description{
Create flagged cell element
}
