% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{Game}
\alias{Game}
\title{R6 Class Representing a Game}
\description{
A game consists of a board and a player, it also tracks the number of turns and is used to play the game.
}
\details{
Call the play function to play out a game, or call turn to step through a game
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-turn}{\code{Game$turn()}}
\item \href{#method-play}{\code{Game$play()}}
\item \href{#method-new}{\code{Game$new()}}
\item \href{#method-print}{\code{Game$print()}}
\item \href{#method-clone}{\code{Game$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-turn"></a>}}
\if{latex}{\out{\hypertarget{method-turn}{}}}
\subsection{Method \code{turn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$turn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-play"></a>}}
\if{latex}{\out{\hypertarget{method-play}{}}}
\subsection{Method \code{play()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$play()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$new(board = Board$new(mines = 20), player = Player$new())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
