% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{gameStatus}
\alias{gameStatus}
\title{Given a game and a state return the status of the game}
\usage{
gameStatus(game, state)
}
\arguments{
\item{game}{the game}

\item{state}{the state of the game}
}
\description{
Given a game and a state return the status of the game
}
