% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{gridUI}
\alias{gridUI}
\title{Helper function to create gridUI}
\usage{
gridUI(inputId, nrow, ncol, status, fun)
}
\arguments{
\item{inputId}{the id of the input}

\item{nrow}{number of row}

\item{ncol}{number of column}

\item{status}{status of the game}

\item{fun}{function that take i, j coordinates and return cell element}
}
\description{
Helper function to create gridUI
}
