% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{hiddenCellUI}
\alias{hiddenCellUI}
\title{Create hidden cell element}
\usage{
hiddenCellUI(i, j)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}
}
\description{
Create hidden cell element
}
