% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{mineCellUI}
\alias{mineCellUI}
\title{Create mine cell element}
\usage{
mineCellUI(i, j)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}
}
\description{
Create mine cell element
}
