% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{minesweeperServer}
\alias{minesweeperServer}
\title{Create the minesweeper backend}
\usage{
minesweeperServer(
  id,
  reactiveStartGame,
  reactiveNrow,
  reactiveNcol,
  reactiveNmines
)
}
\arguments{
\item{id}{the id of the minesweeper module}

\item{reactiveStartGame}{reactive value that trigger a new game}

\item{reactiveNrow}{reactive value that give the number of row}

\item{reactiveNcol}{reactive value that give the number of column}

\item{reactiveNmines}{reactive value that give the number of mines}
}
\description{
Create the minesweeper backend
}
