% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{minesweeperUI}
\alias{minesweeperUI}
\title{Create the full minesweeper ui}
\usage{
minesweeperUI(id)
}
\arguments{
\item{id}{the id of the minesweeper module}
}
\description{
Create the full minesweeper ui
}
