% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{nearbyCoords}
\alias{nearbyCoords}
\title{Given i, j return a vector containing all nearby coordinates}
\usage{
nearbyCoords(i, j, nrow, ncol)
}
\arguments{
\item{i}{the row coordinate}

\item{j}{the column coordinate}

\item{nrow}{the number of rows of the grid}

\item{ncol}{the number of columns of the grid}
}
\description{
Given i, j return a vector containing all nearby coordinates
}
