% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{ongoingGridUI}
\alias{ongoingGridUI}
\title{Create grid element of ongoing game}
\usage{
ongoingGridUI(inputId, game, state)
}
\arguments{
\item{inputId}{the input id}

\item{game}{the game to display}

\item{state}{the current state of the game}
}
\description{
Create grid element of ongoing game
}
