% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_minesweeper.R
\name{play_minesweeper}
\alias{play_minesweeper}
\title{Play Minesweeper}
\usage{
play_minesweeper(
  nrow = 16,
  ncol = 30,
  mine_count = mine_density * nrow * ncol,
  mine_density = 0.20625
)
}
\arguments{
\item{nrow, ncol}{dimensions of the minesweeper board}

\item{mine_count}{number of mines to sweep}

\item{mine_density}{proportion of cells that conceal a mine}
}
\value{
Object of class "minesweeper_recording" to pass to
\code{\link[=replay_minesweeper]{replay_minesweeper()}}, invisibly.
}
\description{
Start a minesweeper game in the current graphics device.
}
\details{
The current graphics device must support event handling
(see \code{\link[grDevices:getGraphicsEvent]{grDevices::getGraphicsEvent()}}). On the \code{windows()} device,
the timer will only update upon mouse events as \code{onIdle} is not supported.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dev.new(noRStudioGD = TRUE)
try(recording <- play_minesweeper())
\dontshow{\}) # examplesIf}
}
