% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{Player}
\alias{Player}
\title{R6 Class Representing a Player}
\description{
This class represents a very basic player that plays randomly and will never feel confident enough to submit a solution
}
\details{
More sophisticated players should inherit this class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_submission}{\code{Player$get_submission()}}
\item \href{#method-get_next_question}{\code{Player$get_next_question()}}
\item \href{#method-recieve_info}{\code{Player$recieve_info()}}
\item \href{#method-submit}{\code{Player$submit()}}
\item \href{#method-new}{\code{Player$new()}}
\item \href{#method-print}{\code{Player$print()}}
\item \href{#method-clone}{\code{Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_submission"></a>}}
\if{latex}{\out{\hypertarget{method-get_submission}{}}}
\subsection{Method \code{get_submission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$get_submission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_next_question"></a>}}
\if{latex}{\out{\hypertarget{method-get_next_question}{}}}
\subsection{Method \code{get_next_question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$get_next_question()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recieve_info"></a>}}
\if{latex}{\out{\hypertarget{method-recieve_info}{}}}
\subsection{Method \code{recieve_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$recieve_info(question, info)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$submit(candidate_minefield)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$new(height = 10, width = 10, mines = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
