% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay_minesweeper.R
\name{replay_minesweeper}
\alias{replay_minesweeper}
\title{Replay Minesweeper}
\usage{
replay_minesweeper(recording)
}
\arguments{
\item{recording}{object of class "minesweeper_recording" returned by
\code{\link[=play_minesweeper]{play_minesweeper()}}}
}
\value{
\code{recording}, invisibly.
}
\description{
Replay a recorded game of minesweeper in the current graphics device.
}
