% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{revealedCellUI}
\alias{revealedCellUI}
\title{Create revealed cell element}
\usage{
revealedCellUI(i, j, nearby_mines)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}

\item{nearby_mines}{number of nearby mines}
}
\description{
Create revealed cell element
}
