% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{unflagCell}
\alias{unflagCell}
\title{Given a state and i, j return a new state with the cell i, j unflagged}
\usage{
unflagCell(state, i, j)
}
\arguments{
\item{state}{the state of the game}

\item{i}{the row coordinate}

\item{j}{the column coordinate}
}
\description{
Given a state and i, j return a new state with the cell i, j unflagged
}
