% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{wrongCellUI}
\alias{wrongCellUI}
\title{Create wrong cell element}
\usage{
wrongCellUI(i, j)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}
}
\description{
Create wrong cell element
}
