% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-module.R
\name{wrongFlagCellUI}
\alias{wrongFlagCellUI}
\title{Create wrongly flagged cell element}
\usage{
wrongFlagCellUI(i, j)
}
\arguments{
\item{i}{row of the cell}

\item{j}{column of the cell}
}
\description{
Create wrongly flagged cell element
}
