% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-package.R
\docType{package}
\name{minesweeper-package}
\alias{minesweeper}
\alias{minesweeper-package}
\title{minesweeper: Play Minesweeper}
\description{
Play, record and replay games of minesweeper using a graphics device that supports event handling.
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
