% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper-package.R
\docType{package}
\name{minesweeper-package}
\alias{minesweeper}
\alias{minesweeper-package}
\title{minesweeper: Play Minesweeper}
\description{
Play and record games of minesweeper using a graphics device that supports event handling. Replay recorded games and save GIF animations of them. Based on classic minesweeper as detailed by Crow P. (1997) \url{https://minesweepergame.com/math/a-mathematical-introduction-to-the-game-of-minesweeper-1997.pdf}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/minesweeper}
  \item Report bugs at \url{https://github.com/hrryt/minesweeper/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk} [copyright holder]

}
\keyword{internal}
