% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minesweeper-package.R
\docType{package}
\name{Minesweeper-package}
\alias{Minesweeper}
\alias{Minesweeper-package}
\title{Minesweeper: A simple Minesweeper implementation in R with shiny}
\description{
This package provides a simple implementation of the Minesweeper game with a graphical interface made with shiny.
}
\author{
\strong{Maintainer}: Lilou Zulewski \email{114218057+lilouzulewski@users.noreply.github.com}

Authors:
\itemize{
  \item Mathieu Le Séac'h \email{113145496+mle-seach@users.noreply.github.com}
}

}
\keyword{internal}
